+++ src/ocean/core/Exception.d
@@ -121,3 +121,224 @@ unittest
         ++counter;
     assert (counter == 2);
 }
+
+/******************************************************************************
+
+    Common code to implement reusable exception semantics - one that has
+    mutable message buffer where message data gets stored to
+
+******************************************************************************/
+
+public template ReusableExceptionImplementation()
+{
+    import tango.text.convert.Integer;
+    static import ocean.core.Array;
+
+    static assert (is(typeof(this) : Exception));
+
+    /**************************************************************************
+
+        Fields used instead of `msg` for mutable messages. `Exception.msg`
+        has `istring` type thus can't be overwritten with new data
+
+    ***************************************************************************/
+
+    protected mstring reused_msg;
+
+    /***************************************************************************
+
+        Constructs exception object with mutable buffer pre-allocated to length
+        `size` and other fields kept invalid.
+
+        Params:
+            size = initial size/length of mutable message buffer
+
+    ***************************************************************************/
+
+    this (size_t size = 128)
+    {
+        this.reused_msg.length = size;
+        super(null);
+    }
+
+    /***************************************************************************
+
+        Sets exception information for this instance.
+
+        Params:
+            msg  = exception message
+
+        Returns:
+            this instance
+
+    ***************************************************************************/
+
+    public typeof (this) set ( cstring msg, istring file = __FILE__,
+        long line = __LINE__ )
+    {
+        ocean.core.Array.copy(this.reused_msg, msg);
+        this.msg  = null;
+        this.file = file;
+        this.line = line;
+        return this;
+    }
+
+   /***************************************************************************
+
+        Throws this instance if ok is false, 0 or null.
+
+        Params:
+            ok   = condition to enforce
+            msg  = exception message
+
+        Throws:
+            this instance if ok is false, 0 or null.
+
+    ***************************************************************************/
+
+    public void enforce ( T ) ( T ok, lazy cstring msg, istring file = __FILE__,
+        long line = __LINE__ )
+    {
+        if (!ok)
+            throw this.set(msg, file, line);
+    }
+
+    version (D_Version2)
+    {
+        /**********************************************************************
+
+            Params:
+                sink = delegate that will be called with parts of currently
+                    active exception message
+
+        ***********************************************************************/
+
+        mixin(`
+        public override void toString(scope void delegate(in char[]) sink) const
+        {
+            sink(this.msg is null ? this.reused_msg : this.msg);
+        }
+        
+        alias toString = super.toString;
+        `);
+    }
+    else
+    {
+        /**********************************************************************
+
+            Returns:
+                currently active exception message
+
+        **********************************************************************/
+
+        public override istring toString()
+        {
+            return this.msg is null ? this.reused_msg : this.msg;
+        }
+    }
+
+    /**************************************************************************
+
+        Appends new substring to mutable exception message
+
+        Intended to be used in hosts that do dynamic formatting of the
+        error message and want to avoid repeating allocation with help
+        of ReusableException semantics
+
+        Leaves other fields untouched
+
+        Params:
+            msg = string to append to the message
+        
+        Returns:
+            this instance
+
+    ***************************************************************************/
+
+    public typeof (this) append ( istring msg )
+    {
+        ocean.core.Array.append(this.reused_msg, msg);
+        return this;
+    }
+
+    /**************************************************************************
+
+        Appends an integer to mutable exception message
+
+        Does not cause any appenditional allocations
+
+        Params:
+            num = number to be formatted
+            hex = optional, indicates that value needs to be formatted as hex
+
+        Returns:
+            this instance
+
+    **************************************************************************/
+
+    public typeof (this) append ( long num, bool hex = false )
+    {
+        char[long.max.stringof.length + 1] buff;
+        if (hex)
+        {
+            ocean.core.Array.append(this.reused_msg, "0x");
+            ocean.core.Array.append(this.reused_msg, format (buff, num, "X"));
+        }
+        else
+            ocean.core.Array.append(this.reused_msg, format (buff, num));
+        return this;
+    }
+}
+
+///
+unittest
+{
+    auto e = new SomeReusableException(100);
+
+    e.set("message");
+    assert (e.toString() == "message");
+    auto old_ptr = e.reused_msg.ptr;
+
+    try
+    {
+        enforce(e, false, "immutable");
+        assert (false);
+    }
+    catch (SomeReusableException) { }
+    assert (e.toString() == "immutable");
+
+    try
+    {
+        e.enforce(false, "longer message");
+    }
+    catch (SomeReusableException) { }
+    assert (e.toString() == "longer message");
+    assert (old_ptr is e.reused_msg.ptr);
+
+    try
+    {
+        e.badName("NAME", 42);
+    }
+    catch (SomeReusableException) { }
+    assert (e.toString() == "Wrong name (NAME) 0x2A 42");
+    assert (old_ptr is e.reused_msg.ptr);
+}
+
+version (UnitTest)
+{
+    private class SomeReusableException : Exception
+    {
+        void badName(istring name, uint id)
+        {
+            this.set("Wrong name (")
+                .append(name)
+                .append(") ")
+                .append(id, true)
+                .append(" ")
+                .append(id);
+            throw this;
+        }
+
+        mixin ReusableExceptionImplementation!();
+    }
+}