+++ src/ocean/util/ReusableException.d
@@ -8,6 +8,8 @@
 
 module ocean.util.ReusableException;
 
+import tango.transition;
+
 /*******************************************************************************
 
     Enhances Exception with additional mutable message buffer that gets reused
@@ -22,18 +24,7 @@ module ocean.util.ReusableException;
 
 class ReusableException : Exception
 {
-    import tango.transition;
-
-    import ocean.core.Array : copy;
-
-    /**************************************************************************
-
-        Fields used instead of `msg` for mutable messages. `Exception.msg`
-        has `istring` type thus can't be overwritten with new data
-
-    ***************************************************************************/
-
-    private mstring reused_msg;
+    mixin ReusableExceptionImplementation!();
 
     /**************************************************************************
 
@@ -48,7 +39,7 @@ class ReusableException : Exception
         Throws this instance if ok is false, 0 or null.
 
         Params:
-            ok   = condition to assert
+            ok   = condition to enforce
             msg  = exception message
             file = source code file
             line = source code line
@@ -57,21 +48,6 @@ class ReusableException : Exception
             this instance if ok is false, 0 or null.
 
     ***************************************************************************/
-
-    public void enforce ( T ) ( T ok, lazy cstring msg, istring file = __FILE__,
-        long line = __LINE__ )
-    {
-        static if (is (T : typeof (null)))
-        {
-            bool err = ok is null;
-        }
-        else
-        {
-            bool err = !ok;
-        }
-
-        if (err) throw this.opCall(msg, file, line);
-    }
 
     deprecated("Use ReusableException.enforce instead")
     public void assertEx ( T ) ( T ok, cstring msg, istring file, long line )
@@ -84,57 +60,20 @@ class ReusableException : Exception
         Sets exception information for this instance.
 
         Params:
-            ok   = condition to assert
             msg  = exception message
             file = source code file
+            line = source code line
 
         Returns:
             this instance
 
     ***************************************************************************/
 
+    deprecated("Use ReusableException.set instead")
     public typeof (this) opCall ( lazy cstring msg, istring file = __FILE__,
         long line = __LINE__ )
     {
-        this.reused_msg.copy(msg);
-        super.msg  = null;
-        super.file = file;
-        super.line = line;
-        return this;
-    }
-
-    version (D_Version2)
-    {
-        /**********************************************************************
-
-            Params:
-                sink = delegate that will be called with parts of currently
-                    active exception message
-
-        ***********************************************************************/
-
-        mixin(`
-        public override void toString(scope void delegate(in char[]) sink) const
-        {
-            sink(this.msg is null ? this.reused_msg : this.msg);
-        }
-        
-        alias toString = super.toString;
-        `);
-    }
-    else
-    {
-        /**********************************************************************
-
-            Returns:
-                currently active exception message
-
-        **********************************************************************/
-
-        public override istring toString()
-        {
-            return this.msg is null ? this.reused_msg : this.msg;
-        }
+        return this.set(msg, file, line);
     }
 }
 
@@ -142,22 +81,6 @@ version (UnitTest)
 {
     import ocean.core.Test;
     import ocean.core.Exception;
-}
-
-unittest
-{
-    auto e = new ReusableException;
-
-    e("message");
-    test!("==")(e.toString(), "message");
-    auto old_ptr = e.reused_msg.ptr;
-
-    testThrown!(ReusableException)(enforce(e, false, "immutable"));
-    test!("==")(e.toString(), "immutable");
-
-    testThrown!(ReusableException)(e.enforce(false, "longer message"));
-    test!("==")(e.toString(), "longer message");
-    test!("is")(old_ptr, e.reused_msg.ptr);
 }
 
 unittest
@@ -172,6 +95,6 @@ unittest
     }
     catch (ReusableException ex)
     {
-        ex("Failed to parse number", __FILE__, __LINE__);
+        ex.set("Failed to parse number", __FILE__, __LINE__);
     }
 }